Working Adding UI Elements - UI can now add static UI elements and strings - Todo: robot id in packets is currently hard code due to issue with it being 0 when assigning robot id in ref_opp

+++ core/util/communications/ui/ui_interface.h
@@ -9,7 +9,7 @@
 #include "ui_types.h"
 #include <functional>
 
-extern int ui_self_id;
+extern uint16_t ui_self_id;
 extern std::function<void(uint8_t*, uint16_t )> send_packet_func;
 
 void print_message(const uint8_t* message, int length);
